% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{add.pheno.bars}
\alias{add.pheno.bars}
\title{Add phenotype barplots}
\usage{
add.pheno.bars(
  plot.df,
  bar.mids,
  bar.hex,
  control.sep,
  horiz = TRUE,
  color.by.sig = TRUE,
  legend.on.bars = FALSE
)
}
\arguments{
\item{plot.df}{Plotting dataframe. See \code{\link[=barplot.by.phenotype]{barplot.by.phenotype()}}.}

\item{bar.mids}{Midpoints for all bars (controls first)}

\item{bar.hex}{Bar width}

\item{control.sep}{Separation constant between cases & controls}

\item{horiz}{Should the bars be plotted horizontally? [default: TRUE]}

\item{color.by.sig}{Should the case bars be shaded by significance? [default: TRUE]}

\item{legend.on.bars}{Should a legend be written on the bars? [default: FALSE]}
}
\description{
Add stacked barplots by phenotype to an existing plot
}
\seealso{
\code{\link[=barplot.by.phenotype]{barplot.by.phenotype()}}
}
