% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{barplot.by.phenotype}
\alias{barplot.by.phenotype}
\title{Barplot of values per cancer type}
\usage{
\method{barplot}{by.phenotype}(
  plot.df,
  bar.hex = 0.8,
  add.top.axis = TRUE,
  top.axis.units = NULL,
  title = "Value",
  parmar = c(0.2, 4.1, 2.1, 4)
)
}
\arguments{
\item{plot.df}{Data.frame of values to plot. See \code{Details}.}

\item{bar.hex}{Relative width for each bar [default: 0.85]}

\item{add.top.axis}{Should the top axis be added? [default: TRUE]}

\item{top.axis.units}{Specify units for top axis. Options are NULL (for
numeric) or "percent" [default: NULL]}

\item{title}{Custom title for top axis}

\item{parmar}{Value of \code{mar} passed to \code{par()}}
}
\description{
Generate a barplot of one value for each cancer type
}
\details{
\code{plot.df} is expected to adhere to the following specifications:
\itemize{
\item One row per cancer type and one row for controls
\item Row names indicate cancer type (or "control")
\item Columns ordered as follows: (1) value to plot, (2) lower CI bound,
(3) upper CI bound, (4) P-value
}
}
