% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.R
\name{calc.genomic.imbalance}
\alias{calc.genomic.imbalance}
\title{Get net genomic imbalance for each row in an SV BED}
\usage{
calc.genomic.imbalance(bed, cnv.type = c("DEL", "DUP"))
}
\arguments{
\item{bed}{SV BED file loaded by \link{load.sv.bed}}

\item{cnv.type}{Restrict computation to deletion or duplication
[default: absolute sum of all gains and losses]}
}
\value{
numeric vector of length \code{nrow(bed)}
}
\description{
For each variant in an SV BED, compute the sum total of bases involved
in deletion or duplication
}
\details{
Recognized options for \code{cnv.type} include "DEL" and "DUP"
}
\seealso{
\code{\link[=load.sv.bed]{load.sv.bed()}}
}
