% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{cowplot}
\alias{cowplot}
\title{Cowplot}
\usage{
cowplot(
  data,
  names = NULL,
  cow.overlap = 0.35,
  xlims = NULL,
  x.axis = TRUE,
  fill = "grey70",
  border = "grey35",
  border.lwd = 2,
  parmar = c(2.5, 3, 0.25, 0.25)
)
}
\arguments{
\item{data}{list of \link{density} objects to plot}

\item{names}{optional list of names for Y axis [default: take names from data]}

\item{cow.overlap}{relative fraction of overlap bewtween adjacent cows [default: 0.35]}

\item{xlims}{custom X axis limits}

\item{fill}{vector of polygon fill colors [default: "grey70"]}

\item{border}{vector of cow border colors [default: "grey35"]}

\item{border.lwd}{line width for cow borders [default: 2]}

\item{parmar}{vector of values passed to par(mar)}

\item{ylims}{custom Y axis limits}
}
\description{
Generate a cowplot using base R syntax
}
