% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{doublewide.barplot.by.phenotype}
\alias{doublewide.barplot.by.phenotype}
\title{Double-wide barpolot of values per cancer type}
\usage{
doublewide.barplot.by.phenotype(
  plot.df.l,
  plot.df.r,
  bar.wex = 0.5,
  case.control.sep = 0.375,
  group.spacer = 0.5,
  color.by.sig = TRUE,
  add.left.axis = TRUE,
  left.axis.units = NULL,
  title = "Value",
  y.title.line = 0.35,
  orient.cases = "left",
  label.l = NULL,
  label.r = NULL,
  group.label.cex = 1,
  parmar = c(1.15, 2.25, 0.75, 0.25)
)
}
\arguments{
\item{plot.df.l}{Data.frame of values to plot on the left. See \code{\link[=barplot.by.phenotype]{barplot.by.phenotype()}}.}

\item{plot.df.r}{Data.frame of values to plot on the right. See \code{\link[=barplot.by.phenotype]{barplot.by.phenotype()}}.}

\item{bar.wex}{Relative width for each bar [default: 0.5]}

\item{case.control.sep}{Relative separation between overlapping case
and control bars [default: 0.375]}

\item{group.spacer}{Distance between left and right barplot groups [default: 0.5]}

\item{color.by.sig}{Should bars be shaded differently by significance
level? [default: TRUE]}

\item{add.left.axis}{Should the left axis be added? [default: TRUE]}

\item{left.axis.units}{Specify units for left axis. Options are NULL (for
numeric) or "percent" [default: NULL]}

\item{title}{Custom title for left axis}

\item{y.title.line}{\code{title.line} parameter passed to \code{\link[RLCtools:clean.axis]{RLCtools::clean.axis()}}
for left axis [default: 0.35]}

\item{orient.cases}{Should cases be plotted to the \code{left} or \code{right} of controls? [default: "left"]}

\item{label.l}{Optional label for left set of bars.}

\item{label.r}{Optional label for right set of bars.}

\item{group.label.cex}{Character expansion used for \code{label.l} and \code{label.r} [default: 1]}

\item{parmar}{Value of \code{mar} passed to \code{par()}}
}
\description{
Generate two side-by-side sets of barplots of two values for each cancer type
}
\details{
This function is effectively a double-wide set of \code{\link[=barplot.by.phenotype]{barplot.by.phenotype()}}.
As such, please refer to \code{\link[=barplot.by.phenotype]{barplot.by.phenotype()}} for more information.
}
\seealso{
\code{\link[=barplot.by.phenotype]{barplot.by.phenotype()}}
}
