% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get.complete.trios}
\alias{get.complete.trios}
\title{Get trios from metadata}
\usage{
get.complete.trios(meta)
}
\arguments{
\item{meta}{Sample metadata loaded by \link{load.sample.metadata}}
}
\value{
data.frame
}
\description{
Infer complete trios from sample metadata
}
\details{
The function will attempt to label mothers and fathers based on chrX
ploidy. This may not always match expectations for trios with data derived from
cell lines (e.g., 1000 Genomes LCLs) given the potential for sex chromosome
aneuploidies arising in culture.
}
