% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{get.eligible.samples}
\alias{get.eligible.samples}
\title{Get eligible samples}
\usage{
get.eligible.samples(meta, cancer)
}
\arguments{
\item{meta}{Sample metadata as loaded by \link{load.sample.metadata}}

\item{cancer}{Cancer code; one of "pancan", "OS", "NBL", or "EWS"}
}
\value{
List of character vectors of sample IDs for cases and controls
}
\description{
Determine the sample IDs eligible for a specified case:control comparison
}
\details{
If \code{meta} contains columns for cancer-specific control designation,
those will be used; otherwise all eligible controls will be returned
}
