% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{get.phenotype.vector}
\alias{get.phenotype.vector}
\title{Build a phenotype indicator vector}
\usage{
get.phenotype.vector(case.ids, control.ids)
}
\arguments{
\item{case.ids}{Vector of case IDs}

\item{control.ids}{Vector of control IDs}
}
\value{
Numeric vector
}
\description{
Construct a one-hot encoding for disease status for a set of cases and controls
}
