% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{hex2grey}
\alias{hex2grey}
\title{Convert colors to greyscale}
\usage{
hex2grey(in.colors)
}
\arguments{
\item{in.colors}{}
}
\value{
character vector of HEX color
}
\description{
Convert one or more HEX-encoded color(s) to its greyscale HEX equivalent(s)
}
\examples{
hex2grey(c("#FF0000", "#00008B", "#ffffe0"))

}
