% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.burden.sumstats}
\alias{load.burden.sumstats}
\title{Load burden summary statistics}
\usage{
load.burden.sumstats(ss.tsv)
}
\arguments{
\item{ss.tsv}{Burden summary statistics generated by generated by global_burden_tests.R}
}
\value{
data.frame
}
\description{
Load global SV burden summary statistics
}
