% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{load.constants}
\alias{load.constants}
\title{Load Study Constants}
\usage{
load.constants(subset, envir = .GlobalEnv)
}
\arguments{
\item{envir}{Environment passed to \link[base:assign]{base::assign} [default: .GlobalEnv]}

\item{susbet}{Vector of constant groups to load See \code{Details} for options.}
}
\description{
Load a subset of constants for germline SV analyses in pediatric cancers
}
\details{
Recognized values for \code{subset} include:
\itemize{
\item \code{colors} : all color palettes used
\item \code{scales} : all scales and scale labels
\item \code{names} : names of various variables
\item \code{all} : load all constants
}
}
\examples{
# Load list of color palettes
get.constants("colors");

# Load scales and names colors
get.constants(c("scales", "names"))

}
\seealso{
\link[base:assign]{base::assign}
}
