% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.kinship.metrics}
\alias{load.kinship.metrics}
\title{Load relatedness metrics}
\usage{
load.kinship.metrics(tsv.in)
}
\arguments{
\item{tsv.in}{Path to input .tsv}
}
\value{
data.frame
}
\description{
Load a .tsv of PC-Relate output metrics for pairs of samples
}
