% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.pc.matrix}
\alias{load.pc.matrix}
\title{Load a principal components matrix}
\usage{
load.pc.matrix(tsv.in, keep.n = 100000000000)
}
\arguments{
\item{keep.n}{Keep top N PCs [default: keep all PCs]}

\item{tsv_in}{Path to input PC matrix .tsv}
}
\value{
data.frame
}
\description{
Load a .tsv of genetic principal components (columns) for a set of samples (rows)
}
