% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.sample.metadata}
\alias{load.sample.metadata}
\title{Load a .tsv of sample metadata}
\usage{
load.sample.metadata(tsv.in, keep.samples = NULL)
}
\arguments{
\item{tsv.in}{Path to input .tsv}

\item{keep.samples}{Vector of sample IDs to retain \link{default: keep all samples}}
}
\value{
data.frame
}
\description{
Load a .tsv of sample metadata
}
\details{
Row names in output dataframe are sample IDs
}
