% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.sv.bed}
\alias{load.sv.bed}
\title{Load a .bed for an SV callset}
\usage{
load.sv.bed(
  bed.in,
  keep.coords = TRUE,
  pass.only = TRUE,
  split.coding = TRUE,
  split.noncoding = FALSE
)
}
\arguments{
\item{bed.in}{Path to input .bed}

\item{pass.only}{Should only PASS variants be included? \link{Default: TRUE}}

\item{split.coding}{Should coding consequence annotations be split from
comma-delimited strings to character vectors? \link{Default: TRUE}}

\item{split.noncoding}{Should noncoding consequence annotations be split from
comma-delimited strings to character vectors? \link{Default: FALSE}}

\item{keep.coordinates}{Should coordinates be retained? \link{Default: TRUE}}
}
\value{
data.frame
}
\description{
Load a .bed for an SV callset
}
\details{
Row names in output dataframe are variant IDs
}
