% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan plot}
\usage{
manhattan(
  stats,
  p.column,
  chrom.colors,
  gw.sig,
  pt.cex = 0.3,
  title = NULL,
  parmar = c(2, 2.25, 0.25, 0.3)
)
}
\arguments{
\item{stats}{data.frame with at least two columns: "pos" and \code{p.column}}

\item{chrom.colors}{Named vector of point colors for each chromosome}

\item{gw.sig}{P-value threshold for genome-wide significance}

\item{pt.cex}{Value of \code{cex} passed to \code{\link[=points]{points()}} [default: 0.3]}

\item{title}{(Optional) title}

\item{parmar}{Value of \code{mar} passed to \code{\link[=par]{par()}}}

\item{p.columns}{Name of column containing P values}
}
\description{
Plot genome-wide summary statistics in the "Manhattan" style
}
