% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{pc.scatterplot}
\alias{pc.scatterplot}
\title{Plot principal components}
\usage{
pc.scatterplot(
  pcs,
  pc.X,
  pc.Y,
  colors,
  title = NULL,
  x.label.line = NULL,
  x.title = NULL,
  x.title.line = 0.5,
  xlims = NULL,
  y.label.line = NULL,
  y.title = NULL,
  y.title.line = 0.5,
  ylims = NULL,
  legend.vals = NULL,
  legend.labels = NULL,
  cex = 0.3,
  parmar = c(2.5, 2.5, 1, 1)
)
}
\arguments{
\item{pcs}{Matrix of principal components as loaded by \link{load.pc.matrix}}

\item{pc.X}{PC to be plotted on the X-axis (see \code{Details})}

\item{pc.Y}{PC to be plotted on the Y-axis (see \code{Details})}

\item{colors}{Vector of colors. Must be supplied in the same order as rows in \code{pcs}}

\item{title}{Main title for plot [default: NULL]}

\item{x.label.line}{Line for X-axis labels (\code{label.line} parameter for \link[RLCtools:clean.axis]{RLCtools::clean.axis})}

\item{x.title}{Title for X-axis}

\item{x.title.line}{Line for X-axis title (\code{title.line} parameter for \link[RLCtools:clean.axis]{RLCtools::clean.axis})}

\item{xlims}{Limits for X-axis}

\item{y.label.line}{Line for X-axis labels (\code{label.line} parameter for \link[RLCtools:clean.axis]{RLCtools::clean.axis})}

\item{y.title}{Title for Y-axis}

\item{y.title.line}{Line for axis titles (\code{title.line} parameter for \link[RLCtools:clean.axis]{RLCtools::clean.axis})}

\item{ylims}{Limits for Y-axis}

\item{legend.vals}{Named vector mapping category names to colors [default: NULL]}

\item{legend.labels}{Optional vector to overwrite names of \code{legend.vals}}

\item{cex}{Character expansion factor for individual points [default: 0.3]}

\item{parmar}{Numeric vector of values to pass to \code{par(mar)}}
}
\description{
Generate a scatterplot of principal components with options for coloring
}
\details{
\code{pc.X} and \code{pc.Y} can be specified as numeric column indexes (e.g., \code{1}),
or can be specified by column name (e.g., \code{PC1}).

If \code{legend.vals} is not provided (or is \code{NULL}), no legend will be added.
}
