% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{plot.qq}
\alias{plot.qq}
\title{Quantile-quantile plot}
\usage{
\method{plot}{qq}(
  pvals,
  cutoff = NULL,
  do.fdr = TRUE,
  fdr.cutoff = 0.01,
  print.stats = FALSE,
  title = NULL,
  title.line = 0,
  xmax = NULL,
  ymax = NULL,
  label.cex = 1,
  pt.color = "grey35",
  pt.cex = 0.35,
  fdr.color = "grey5",
  fdr.cex = 0.7,
  parmar = c(2.25, 2.5, 0.25, 0.25)
)
}
\arguments{
\item{pvals}{Numeric vector of untransformed P-values}

\item{cutoff}{P-value threshold for significance}

\item{do.fdr}{Should points meeting \code{fdr.cutoff} be accented? [default: TRUE]}

\item{fdr.cutoff}{Cutoff for highlighting FDR-significant points [default: 0.01]}

\item{print.stats}{Should genomic inflation statistic be printed on plot? [default: FALSE]}

\item{title}{(Optional) Plot title}

\item{title.line}{Line for \code{title} [default: 0]}

\item{xmax}{Maximum X-value to plot [default: include all points]}

\item{ymax}{Maximum Y-value to plot [default: smallest P-value]}

\item{label.cex}{Scaling factor for label text [default: 1]}

\item{pt.color}{Color for all points [default: "grey35"]}

\item{pt.cex}{Scaling factor for all points [default: 0.35]}

\item{fdr.color}{Color for FDR-significant points [default: "grey5"]}

\item{fdr.cex}{Scaling factor for FDR-significant points [default: 0.7]}

\item{parmar}{Value of \code{mar} passed to \code{par()}}
}
\value{
None
}
\description{
Generate a Q-Q plot of association stats vs. uniform null
}
