% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{prep.plot.area}
\alias{prep.plot.area}
\title{Prepare Plot Area}
\usage{
prep.plot.area(xlims, ylims, parmar, xaxs = "i", yaxs = "i")
}
\arguments{
\item{xlims}{Range of values for X axis}

\item{ylims}{Range of values for Y axis}

\item{parmar}{Margin values passed to par()}

\item{xaxs}{Value of \code{xaxs} passed to plot()}

\item{yaxs}{Value of \code{yaxs} passed to plot()}
}
\description{
Prepare a standardized & formatted plot area
}
\examples{
prep.plot.area(xlims=c(0, 5), ylims=(-10, 10), parmar=rep(3, 4));

}
