% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleDosage.R
\name{query.ad.from.sv.bed}
\alias{query.ad.from.sv.bed}
\title{Query allele dosage matrix using an SV BED file}
\usage{
query.ad.from.sv.bed(ad, bed, action = "verbose")
}
\arguments{
\item{ad}{Allele dosage matrix. See \link{query.ad.matrix}.}

\item{bed}{Data frame of SVs to query as loaded by \link{load.sv.bed}.}

\item{action}{Action to apply to query. See \link{query.ad.matrix}.}
}
\value{
Data.frame or vector depending on value of \code{action}
}
\description{
Helper function to format and quickly execute multiple allele dosage matrix
queries corresponding to the variants present in an input SV BED file
}
\seealso{
\link{query.ad.matrix}, \link{load.sv.bed}
}
