% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleDosage.R
\name{query.ad.matrix}
\alias{query.ad.matrix}
\title{Query allele dosage matrix}
\usage{
query.ad.matrix(
  ad,
  query.regions = NULL,
  query.ids = NULL,
  action = "verbose",
  weights = NULL,
  na.behavior = "threshold",
  na.frac = 0.05,
  keep.samples = NULL
)
}
\arguments{
\item{ad}{Either a path to an allele dosage .bed.gz matrix or a data.frame
from a previous call of \link{query.ad.matrix}}

\item{query.regions}{List of tripartite coordinate vectors to query; see \code{Details}}

\item{query.ids}{Vector of variant IDs to query}

\item{action}{Action to apply to query rows; see \code{Details}}

\item{weights}{Optional named vector of weights for each variant. Variants
not present in this vector will be assigned weight = 1. [default: no weighting]}

\item{na.behavior}{Specify how \code{NA} entries in \code{ad.df} should be treated when
compressing. See \link{compress.ad.matrix} for more information.}

\item{na.frac}{Fraction of \code{NA} entries allowed before failing a sample.
See \link{compress.ad.matrix} for more information.}

\item{keep.samples}{Optional character vector of sample IDs to retain.
[default: keep all samples]}
}
\value{
numeric vector or data.frame, depending on \code{action}
}
\description{
Subsets and optionally compresses an allele dosage matrix
}
\details{
The value for \code{query.regions} is expected to be a list of vectors,
where each vector must either be a single chromosome or have exactly
three elements in the format (chromosome, start, end).

Recognized values for \code{action} include:
\itemize{
\item \code{"verbose"} : return the full query matrix [default]
\item \code{"any"} : return numeric indicator if any of the query rows are non-zero
for each sample
\item \code{"all"} : return numeric indicator if all of the query rows are non-zero
for each sample
\item \code{"count"} : return the count of non-reference variants per sample
irrespective of genotype
\item \code{"sum"} : return the sum of allele dosages for all query rows per sample
}
}
\seealso{
\link{compress.ad.matrix}
}
