% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{saddlepoint.adj}
\alias{saddlepoint.adj}
\title{Z-score saddlepoint approximation}
\usage{
saddlepoint.adj(zscores, alternative = "two.sided")
}
\arguments{
\item{zscores}{numeric vector of Z-scores}

\item{alternative}{alternative hypothesis. Options include "two.sided", "less", or "greater". [default: "two.sided"]}
}
\value{
data.frame with two columns:
\itemize{
\item \verb{$zscores} for corrected Z-scores
\item \verb{$pvalues} for P-values corresponding to corrected Z-scores
}
}
\description{
Apply saddlepoint approximation to vector of Z-scores to generate adjusted P-values
}
