% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{stretch.vector}
\alias{stretch.vector}
\title{Stretch Vector}
\usage{
stretch.vector(values, k)
}
\arguments{
\item{values}{Vector of values to be stretched}

\item{k}{Number of times to duplicate each element in \code{values}}
}
\description{
Expand the length of a vector by repeating (or "stuttering") values
}
\examples{
stretch.vector(values=c(1, 2, 3), k=4)

}
