% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{sv.size.survfit}
\alias{sv.size.survfit}
\title{Survival model of SV size}
\usage{
sv.size.survfit(
  sv.bed,
  meta,
  ad,
  cancer = "pancan",
  use.N.pcs = 5,
  extra.terms = NULL
)
}
\arguments{
\item{sv.bed}{SV bed that was imported with \link{load.sv.bed}}

\item{meta}{Sample metadata that was imported with \link{load.sample.metadata}}

\item{ad}{Either a path to an allele dosage .bed.gz matrix or a data.frame
from a previous call of \link{query.ad.matrix}}

\item{cancer}{Cancer type of interest [default: "pancan"]}

\item{use.N.pcs}{Specify how many principal components should be adjusted in
model [default: 5]}

\item{extra.terms}{Specify if any extra terms should be added to the model.
Named options include:  "batch", "coverage, "insert.size", and "wgd".
Custom terms can be passed using their exact column names in \code{meta}.}
}
\description{
Constructs a Cox proportional hazards model of the largest SV per sample
versus case-control status
}
