% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{svlen.line.plot}
\alias{svlen.line.plot}
\title{Plot values vs. SV length}
\usage{
svlen.line.plot(
  x.svlen,
  y.value,
  colors,
  ci.lower = NULL,
  ci.upper = NULL,
  group.names = NULL,
  step = TRUE,
  lwds = NULL,
  ci.alpha = NULL,
  legend = FALSE,
  legend.names = NULL,
  legend.pos = "topright",
  legend.label.spacing = 0.075,
  title = NULL,
  xlab = NULL,
  xlab.line = 0,
  xlims = log10(c(10000, 1000000)),
  ylab = "Value",
  y.title.line = 1,
  y.axis.units = NULL,
  ylims = NULL,
  x.axis.labels = NULL,
  x.axis.labels.at = NULL,
  x.tck = -0.01,
  parmar = c(2, 3, 1, 1)
)
}
\arguments{
\item{x.svlen}{List of SV length values to be plotted on X axis; one element per stratum}

\item{y.value}{List of values to be plotted on Y axis; one element per stratum}

\item{colors}{Vector of colors for the list elements in \code{x.svlen} and \code{y.value}}

\item{ci.lower}{(Optional) List of lower confidence interval bounds per \code{y.value}; one element per stratum}

\item{ci.upper}{(Optional) List of lower confidence interval bounds per \code{y.value}; one element per stratum}

\item{group.names}{(Optional) group names to assign to each list element  in \code{x.svlen} and \code{y.value}}

\item{step}{Should the line be rendered as a step function (a la Kaplan-Meyer)? [default: TRUE]}

\item{lwds}{(Optional) vector of \code{lwd} for each line [default: 3]}

\item{ci.alpha}{Transparency value \code{alpha} for confidence interval shading [default: 0.15]}

\item{legend}{Should a legend be plotted? [default: FALSE]}

\item{legend.names}{(Optional) mapping of \code{values} to labels for legend}

\item{legend.pos}{(Optional) position for legend. See \code{\link[=legend]{legend()}} [default: "topright"]}

\item{legend.label.spacing}{Minimum vertical spacing between legend labels [default: 0.075]}

\item{title}{(Optional) Title for plot}

\item{xlab}{(Optional) Title for X axis}

\item{xlab.line}{Line parameter for \code{xlab} [default: 1]}

\item{xlims}{(Optional) two-element vector of start and stop values for X-axis, in log10(SVLEN)}

\item{ylab}{(Optional) Title for Y axis}

\item{y.title.line}{Value for \code{title.line} passed to \code{\link[=clean.axis]{clean.axis()}} [default: 1]}

\item{y.axis.units}{Specify units for Y axis. Options are NULL (for
numeric) or "percent" [default: NULL]}

\item{ylims}{(Optional) two-element vector of start and stop values for Y-axis}

\item{x.axis.labels}{(Optional) custom labels for X axis}

\item{x.axis.labels.at}{(Optional) custom positions for X axis labels}

\item{x.tck}{Value of \code{tck} passed to \code{\link[=clean.axis]{clean.axis()}}}

\item{parmar}{Margin values passed to par()}
}
\description{
Generic function to plot one or more (x, y) datasets versus SV length
}
