% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{swarmplot.by.phenotype}
\alias{swarmplot.by.phenotype}
\title{Swarmplot of values per cancer type}
\usage{
swarmplot.by.phenotype(
  plot.vals,
  meta,
  title = NULL,
  title.line = 0,
  title.cex = 1,
  ylims = NULL,
  add.y.axis = TRUE,
  y.axis.title = "Value",
  y.title.line = 0.5,
  y.ticks = NULL,
  y.tick.labels = NULL,
  parse.labels = TRUE,
  cancer.name.xadj = 0.1,
  shorten.cancer.names = FALSE,
  custom.group.names = NULL,
  add.sample.size = FALSE,
  sample.size.mirror.buffer = 0.2,
  scale.swarms = TRUE,
  median.labels = FALSE,
  median.label.nsmall = 1,
  gutter.width = 0.1,
  pt.cex = NULL,
  shade = "main",
  return.swarm.df = FALSE,
  parmar = c(3, 2.5, 0.3, 0.3)
)
}
\arguments{
\item{plot.vals}{Values to plot. See \code{Details} for formatting.}

\item{meta}{Sample metadata loaded using \code{\link[=load.sample.metadata]{load.sample.metadata()}}}

\item{title}{Custom title for top axis}

\item{title.line}{Line for main title on top axis [default: 0]}

\item{title.cex}{\code{cex} parameter for main title [default: 1]}

\item{ylims}{Two-value vector of limits for Y axis. Individual values outside
of these limits will be Winsorized. [default: use full range of all values]}

\item{add.y.axis}{Should the Y \(left\) axis be added? [default: TRUE]}

\item{y.axis.title}{Text title for Y \(left\) axis [default: "Value"]}

\item{y.title.line}{Line for Y \(left\) axis [default: 0.5]}

\item{y.ticks}{\(Optional\) custom tick positions for Y \(left\) axis}

\item{y.tick.labels}{\(Optional\) custom tick labels for Y \(left\) axis}

\item{parse.labels}{Value of \code{parse.labels} passed to \code{\link[=clean.axis]{clean.axis()}} [default: TRUE]}

\item{cancer.name.xadj}{Horizontal adjustment factor for cancer labels on X axis [default: 0.1]}

\item{shorten.cancer.names}{Should cancer names be shortened as much as possible? [default: FALSE]}

\item{custom.group.names}{Manual override of cancer names for groups to be
labeled on the bottom X axis. Must be provided as a named vector mapping cancer
types to desired label.}

\item{add.sample.size}{Should sample size be added to cancer names? [default: FALSE]}

\item{sample.size.mirror.buffer}{Space in user units that cancer label and sample size should be separated [default: 0.2]}

\item{scale.swarms}{Should the width of each swarm be scaled proportional
to the square root of its sample size? [default: TRUE]}

\item{median.labels}{Should text labels for medians per group be plotted? [default: FALSE]}

\item{median.label.nsmall}{Number of decimal places for rounding \code{medial.labels} [default: 1]}

\item{gutter.width}{Relative width of gutter between swarms [default: 0.1]}

\item{pt.cex}{Value of \code{cex} passed to \code{\link[beeswarm:beeswarm]{beeswarm::beeswarm()}}}

\item{shade}{Shade of phenotype color to use [default: 'main']}

\item{return.swarm.df}{Should swarm coordinates be returned? [default: FALSE]}

\item{parmar}{Value of \code{mar} passed to \code{par()}}
}
\description{
Generate a series of beeswarm plots for values per sample separated by cancer type
}
\details{
\code{plot.vals} can be provided in one of two formats:
\enumerate{
\item As a named numeric vector with vector names set to sample IDs
\item As a data.frame with row names set to sample IDs. If provided as a data.frame,
only the first column will be plotted.
}
}
\seealso{
\code{\link[beeswarm:beeswarm]{beeswarm::beeswarm()}}
}
