% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{add.pheno.bars}
\alias{add.pheno.bars}
\title{Add phenotype barplots}
\usage{
add.pheno.bars(
  plot.df,
  bar.mids,
  bar.hex = 0.5,
  control.sep = 0.1875,
  horiz = TRUE,
  color.by.sig = TRUE,
  legend.on.bars = FALSE,
  case.label = "Case",
  control.label = "Control",
  add.pvals = FALSE,
  cancer.types.override = NULL
)
}
\arguments{
\item{plot.df}{Plotting dataframe. See \code{\link[=barplot.by.phenotype]{barplot.by.phenotype()}}.}

\item{bar.mids}{Midpoints for all bars (controls first)}

\item{bar.hex}{Bar width [default: 0.5]}

\item{control.sep}{Separation constant between cases & controls [default: 0.1875]}

\item{horiz}{Should the bars be plotted horizontally? [default: TRUE]}

\item{color.by.sig}{Should the case bars be shaded by significance? [default: TRUE]}

\item{legend.on.bars}{Should a legend be written on the bars? [default: FALSE]}

\item{case.label}{Label for case bars if \code{legend.on.bars} is \code{TRUE} [default: "Case"]}

\item{control.label}{Label for control bars if \code{legend.on.bars} is \code{TRUE} [default: "Control"]}

\item{add.pvals}{Should P values be annotated on the opposite margin?
[default: TRUE]}

\item{cancer.types.override}{Specify order of cancer types in \code{plot.df}
[default: infer from \code{rownames(plot.df)} ]}
}
\description{
Add stacked barplots by phenotype to an existing plot
}
\seealso{
\code{\link[=barplot.by.phenotype]{barplot.by.phenotype()}}
}
