% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.basic.R
\name{barplot.by.phenotype}
\alias{barplot.by.phenotype}
\title{Barplot of values per cancer type}
\usage{
\method{barplot}{by.phenotype}(
  plot.df,
  bar.hex = 0.5,
  case.control.sep = 0.375,
  color.by.sig = TRUE,
  add.top.axis = TRUE,
  top.axis.units = NULL,
  title = "Value",
  orient.cases = "top",
  custom.pheno.labels = NULL,
  legend.on.bars = FALSE,
  parmar = c(0.2, 4.1, 2.1, 4)
)
}
\arguments{
\item{plot.df}{Data.frame of values to plot. See \code{Details}.}

\item{bar.hex}{Relative width for each bar [default: 0.5]}

\item{case.control.sep}{Relative separation between overlapping case and control bars [default: 0.375]}

\item{color.by.sig}{Should bars be shaded differently by significance level? [default: TRUE]}

\item{add.top.axis}{Should the top axis be added? [default: TRUE]}

\item{top.axis.units}{Specify units for top axis. Options are NULL (for
numeric) or "percent" [default: NULL]}

\item{title}{Custom title for top axis}

\item{custom.pheno.labels}{Custom phenotype labels for the groups on the
Y axis, if desired.}

\item{legend.on.bars}{Should "case" and "control" labels be printed on the
largest bars? [default: FALSE]}

\item{parmar}{Value of \code{mar} passed to \code{par()}}
}
\description{
Generate a barplot of one value for each cancer type
}
\details{
\code{plot.df} is expected to adhere to the following specifications:
\itemize{
\item One row per cancer type
\item Row names indicate cancer type (or "control")
\item Columns ordered as follows: (1) case value to plot, (2) case lower CI bound,
(3) case upper CI bound, (4) control value to plot, (5) control lower CI bound,
(6) control upper CI bound, (7) P-value
}
}
