% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{clean.axis}
\alias{clean.axis}
\title{Clean axis}
\usage{
clean.axis(
  side,
  at = NULL,
  labels = NULL,
  title = NULL,
  tck = -0.025,
  cex.axis = 5/6,
  label.line = -0.65,
  title.line = 0.5,
  infinite = FALSE
)
}
\arguments{
\item{side}{Value passed to \code{axis()}. See \code{?axis} for details.}

\item{at}{Positions where axis ticks should be plotted [default: \code{axTicks(side)}]}

\item{labels}{Labels for axis ticks [default: values of \code{at}]}

\item{title}{Axis title}

\item{tck}{Value passed to \code{axis()}. See \code{?axis} for details. [default: -0.025]}

\item{cex.axis}{Value passed to \code{axis()}. See \code{?axis} for details. [default: 5/6]}

\item{label.line}{\code{line} parameter for axis labels [default: -0.65]}

\item{title.line}{\code{line} parameter for axis title [default: 0.5]}

\item{infinite}{Indicator for the axis to be extended infinitely (without ticks) [default: FALSE]}
}
\description{
Print a clean axis using visually pleasing defaults
}
