% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleDosage.R
\name{compress.ad.matrix}
\alias{compress.ad.matrix}
\title{Compress allele dosage matrix}
\usage{
compress.ad.matrix(ad.df, action, weights = NULL)
}
\arguments{
\item{ad.df}{Data.frame of allele dosages loaded by \link{query.ad.matrix}}

\item{action}{Action to apply to query rows; see \code{Details} in \link{query.ad.matrix}}

\item{weights}{Optional named vector of weights for each variant. Variants
not present in this vector will be assigned weight = 1. [default: no weighting]}
}
\value{
numeric vector
}
\description{
Compresses information in an allele dosage matrix
}
\seealso{
\link{query.ad.matrix}
}
