% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.R
\name{filter.bed}
\alias{filter.bed}
\title{Filter an SV BED}
\usage{
filter.bed(
  bed,
  query,
  af.field = "AF",
  ac.field = "AC",
  autosomal = TRUE,
  pass.only = TRUE,
  keep.idx = NULL
)
}
\arguments{
\item{bed}{SV BED file loaded by \link{load.sv.bed}}

\item{query}{Terms to indicate what kind of filtering should be performed. See \code{details}.}

\item{af.field}{BED column to use for allele frequencies [default: AF]}

\item{ac.field}{BED column to use for allele counts [default: AC]}

\item{autosomal}{Keep only autosomal variants [default: TRUE]}

\item{pass.only}{Keep only variants where FILTER is PASS [default: TRUE]}

\item{keep.idx}{Optional vector to specify which row numbers from \code{bed} should
be retained. Can be useful for applying more complex or custom filtering in addition
to the basic functions in \code{query}.}
}
\value{
data.frame
}
\description{
Filter an SV BED based on various categories of interest
}
\details{
The \code{query} argument can be passed as either a character vector or a
single character string. In both cases, the argument should indicate what filter(s)
should be applied to \code{bed}. Recognized filters currently include:
\itemize{
\item \code{rare} : AF < 1\%
\item \code{vrare} : AF < 0.1\%
\item \code{singleton} : AC = 1
\item \code{large} : SVLEN > 1,000,000
\item \code{gene_disruptive} or \code{genes_disrupted} : any SV with predicted LoF, PED, CG, or IED
\item \code{single_gene_disruptive} : as above, but further restricting to SVs impacting just one gene
\item \code{lof} : predicted LoF and/or PED consequences
\item \code{cg} : predicted CG consequence
\item \code{ied} : predicted IED consequence
\item \code{cg_and_ied} : predicted CG and/or IED
}

If \code{query} is provided as a vector, it should contain any of the above terms.
If it is provided as a single character string, it must be period-delimited.
}
