% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.R
\name{filter.bed}
\alias{filter.bed}
\title{Filter an SV BED}
\usage{
\method{filter}{bed}(
  bed,
  query,
  af.field = "POPMAX_AF",
  ac.field = "AC",
  use.gnomad.freqs = TRUE,
  gnomad.max.freq = 0.01,
  gnomad.column = "gnomad_v3.1_sv_POPMAX_AF",
  autosomal = FALSE,
  pass.only = TRUE,
  keep.idx = NULL,
  return.idxs = FALSE
)
}
\arguments{
\item{bed}{SV BED file loaded by \link{load.sv.bed}}

\item{query}{Terms to indicate what kind of filtering should be performed. See \code{details}.}

\item{af.field}{BED column to use for allele frequencies [default: POPMAX_AF]}

\item{ac.field}{BED column to use for allele counts [default: AC]}

\item{use.gnomad.freqs}{Should gnomAD SV AFs also be applied for
frequency filtering? [default: TRUE]}

\item{gnomad.max.freq}{Maximum frequency allowed in gnomAD [default: 0.01]}

\item{gnomad.column}{Specify the column header to use for gnomAD frequency
filtering. Only used if \code{use.gnomad.freqs == TRUE}. [default: "gnomad_v3.1_sv_POPMAX_AF"]}

\item{autosomal}{Keep only autosomal variants [default: FALSE]}

\item{pass.only}{Keep only variants where FILTER is PASS [default: TRUE]}

\item{keep.idx}{Optional vector to specify which row numbers from \code{bed} should
be retained. Can be useful for applying more complex or custom filtering in addition
to the basic functions in \code{query}.}

\item{return.idxs}{Should row indexes be returned instead of a filtered version
of \code{bed}? [default: FALSE]}
}
\value{
data.frame
}
\description{
Filter an SV BED based on various categories of interest
}
\details{
The \code{query} argument can be passed as either a character vector or a
single character string. In both cases, the argument should indicate what filter(s)
should be applied to \code{bed}. Recognized filters currently include:
\itemize{
\item \code{DEL}|\code{DUP}|\code{CNV}|\code{INS}|\code{INV}|\code{CPX}|\code{CTX} : retain only this SV type
\item \code{rare} : AF < 1\%
\item \code{vrare} : AF < 0.1\%
\item \code{singleton} : AC = 1
\item \code{large} : SVLEN > 1,000,000 or SVTYPE = CTX
\item \code{karyotypic} : SVLEN > 5,000,000 or SVTYPE = CTX
\item \code{gene_disruptive} or \code{genes_disrupted} : any SV with predicted LoF, PED, CG, or IED
\item \code{single_gene_disruptive} : as above, but further restricting to SVs impacting just one gene
\item \code{lof} : predicted LoF and/or PED consequences
\item \code{cg} : predicted CG consequence
\item \code{ied} : predicted IED consequence
\item \code{cg_and_ied} : predicted CG and/or IED
}

If \code{query} is provided as a vector, it should contain any of the above terms.
If it is provided as a single character string, it must be period-delimited.

If \code{use.gnomad.freqs} is \code{TRUE} and any of \code{rare}, \code{vrare}, or \code{singleton}
are included in \code{query}, then variants will be further filtered
to be < \code{gnomad.max.freq} according to \code{gnomad.column}
}
