% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.sample.metadata}
\alias{load.sample.metadata}
\title{Load a .tsv of sample metadata}
\usage{
load.sample.metadata(
  tsv.in,
  keep.samples = NULL,
  reassign.parents = TRUE,
  other.keep.columns = NULL
)
}
\arguments{
\item{tsv.in}{Path to input .tsv}

\item{keep.samples}{Either vector of sample IDs to retain or path to
single-column flat text file of sample IDs to retain \link{default: keep all samples}}

\item{reassign.parents}{Assign all parents to have \code{control} disease labels \link{default: TRUE}}

\item{other.keep.columns}{Names of non-standard columns to be retained, if desired}
}
\value{
data.frame
}
\description{
Load a .tsv of sample metadata
}
\details{
Row names in output dataframe are sample IDs
}
