% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.sv.bed}
\alias{load.sv.bed}
\title{Load a .bed for an SV callset}
\usage{
load.sv.bed(
  bed.in,
  keep.coords = TRUE,
  pass.only = TRUE,
  split.coding = TRUE,
  split.noncoding = FALSE,
  split.mcnv.freqs = FALSE,
  drop.vids = NULL,
  drop.cohort.frequencies = c(),
  keep.all.pop.frequencies = FALSE,
  keep.all.sex.frequencies = FALSE
)
}
\arguments{
\item{bed.in}{Path to input .bed}

\item{pass.only}{Should only PASS variants be included? [Default: TRUE]}

\item{split.coding}{Should coding consequence annotations be split from
comma-delimited strings to character vectors? [Default: TRUE]}

\item{split.noncoding}{Should noncoding consequence annotations be split from
comma-delimited strings to character vectors? [Default: FALSE]}

\item{split.mcnv.freqs}{Should mCNV frequency columns be split from
comma-delimited strings to character vectors? [Default: FALSE]}

\item{drop.vids}{Either a vector or a path to a file containing variant IDs
to be excluded drop the input file. [Default: Keep all IDs]}

\item{drop.cohort.frequencies}{A character vector of one or more prefixes
corresponding to cohort names for which cohort-specific frequencies should be
dropped. Useful for reducing in-memory size. Examples include "case_control"
and "trio". [default: Keep all frequencies]}

\item{keep.all.pop.frequencies}{Optional boolean indicator to retain population-
specific frequency info. [default: FALSE]}

\item{keep.all.sex.frequencies}{Optional boolean indicator to retain sex-specific
frequency info. [default: FALSE]}

\item{keep.coordinates}{Should coordinates be retained? [Default: TRUE]}
}
\value{
data.frame
}
\description{
Load a .bed for an SV callset
}
\details{
Row names in output dataframe are variant IDs
}
