% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{pedsv.glm}
\alias{pedsv.glm}
\title{PedSV generic association test}
\usage{
pedsv.glm(meta, X, Y, use.N.pcs = 10, family = gaussian())
}
\arguments{
\item{meta}{Sample metadata loaded with \link{load.sample.metadata}}

\item{X}{Vector of values for primary independent variable. See \link{prep.glm.matrix}.}

\item{Y}{Vector of values for dependent variable. See \link{prep.glm.matrix}.}

\item{use.N.pcs}{Specify how many principal components should be adjusted in
model [default: 10]}

\item{family}{\code{family} parameter passed to \link{glm}}
}
\value{
Named vector of test statsitics corresponding to independent variable
}
\description{
Fit a GLM for user-defined dependent & independent variables while adjusting
for standard sample covariates
}
\seealso{
\link{prep.glm.matrix}
}
