% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{query.ad.matrix}
\alias{query.ad.matrix}
\title{Query allele dosage matrix}
\usage{
query.ad.matrix(ad, query.regions = NULL, query.ids = NULL, action = "verbose")
}
\arguments{
\item{ad}{Either a path to an allele dosage .bed.gz matrix or a data.frame
from a previous call of \link{query.ad.matrix}}

\item{query.regions}{List of tripartite coordinate vectors to query; see \code{Details}}

\item{query.ids}{Vector of variant IDs to query}

\item{action}{Action to apply to query rows; see \code{Details}}
}
\value{
numeric vector or data.frame, depending on \code{action}
}
\description{
Subsets and optionally compresses an allele dosage matrix
}
\details{
The value for \code{query.regions} is expected to be a list of vectors,
where each vector must have three elements of the format (chromosome, start, end).

Recognized values for \code{action} include:
\itemize{
\item \code{"verbose"} : return the full query matrix [default]
\item \code{"any"} : return numeric indicator if any of the query rows are non-zero
for each sample
\item \code{"all"} : return numeric indicator if all of the query rows are non-zero
for each sample
\item \code{"sum"} : return the sum of allele dosages for all query rows per sample
}

Note that
}
