% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{stats2barplotdf}
\alias{stats2barplotdf}
\title{Convert landscape test stats to barplot-compliant data.frame}
\usage{
stats2barplotdf(stats, ci.mode = "normal", order.by.cancer = TRUE)
}
\arguments{
\item{stats}{data.frame with test statistics as output by \code{\link[=pedsv.glm]{pedsv.glm()}}}

\item{ci.mode}{mode for confidence interval calculation [choices: \code{binomial}, \code{normal}; default: \code{normal}]}

\item{order.by.cancer}{order the results by study-wide convention [default: TRUE]}
}
\value{
data.frame
}
\description{
Transform the association summary statistics from a single landscape
case:control burden test into a format expected by \code{\link[=barplot.by.phenotype]{barplot.by.phenotype()}}
}
\seealso{
\code{\link[=barplot.by.phenotype]{barplot.by.phenotype()}}, \code{\link[=pedsv.glm]{pedsv.glm()}}
}
